requirements.txt
mysql_connector_repackaged==0.3.1
pandas==2.2.3
pytest==8.3.4
scipy==1.15.1
SQLAlchemy==2.0.37
